backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
praw==7.8.1
prawcore==2.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-dotenv==1.0.1
pytest-asyncio==0.23.6
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.3.0
uvicorn==0.34.0
vaderSentiment==3.3.2
websocket-client==1.8.0
